+++ src/pages/Calculator/Steps/MixRatios/index.js
@@ -19,13 +19,10 @@ import {
   convertToPercent,
   convertToDecimal,
 } from "./../../../../shared/utils/calculate";
-
 import { updateSteps } from "../../../../features/stepSlice";
-import { NumberTextField } from "./../../../../components/NumberTextField";
-
 import "./../steps.css";
 import SeedsSelectedList from "../../../../components/SeedsSelectedList";
-import { seedsList } from "../../../../shared/data/species";
+import MixRatioSteps from "./form";
 
 const MixRatio = ({ council }) => {
   // themes
@@ -186,7 +183,7 @@ const MixRatio = ({ council }) => {
       </Accordion>
     );
   };
-  const renderAccordianDetail = (data) => {
+  const renderAccordianDetail = (seed) => {
     return (
       <Grid container xs={12}>
         <Grid item xs={6}>
@@ -203,7 +200,7 @@ const MixRatio = ({ council }) => {
             }}
           >
             <Typography>
-              {Math.floor(data.singleSpeciesSeedingRatePLS)}
+              {Math.floor(seed.singleSpeciesSeedingRatePLS)}
             </Typography>
           </Box>
           <Typography>Lbs / Acre</Typography>
@@ -221,7 +218,7 @@ const MixRatio = ({ council }) => {
               borderRadius: "50%",
             }}
           >
-            <Typography>{Math.round(data.mixSeedingRate)}</Typography>
+            <Typography>{Math.round(seed.mixSeedingRate)}</Typography>
           </Box>
           <Typography className="font-15">Lbs / Acre</Typography>
         </Grid>
@@ -239,7 +236,7 @@ const MixRatio = ({ council }) => {
                   borderRadius: "16px",
                 }}
               >
-                <Typography>{Math.floor(data.plantsPerAcre)}</Typography>
+                <Typography>{Math.floor(seed.plantsPerAcre)}</Typography>
               </Box>
               <Typography>Aprox plants per</Typography>
               <Button variant="outlined">Sqft</Button>
@@ -260,7 +257,7 @@ const MixRatio = ({ council }) => {
               borderRadius: "16px",
             }}
           >
-            <Typography>{Math.floor(data.seedsPerAcre)}</Typography>
+            <Typography>{Math.floor(seed.seedsPerAcre)}</Typography>
           </Box>
           <Typography>Seeds per</Typography>
           <Button variant="contained">Sqft</Button>
@@ -270,15 +267,23 @@ const MixRatio = ({ council }) => {
         <Grid item xs={12}>
           <Button
             onClick={(e) => {
-              updateSeed(!data.showSteps, "showSteps", data);
+              updateSeed(!seed.showSteps, "showSteps", seed);
             }}
             variant="outlined"
           >
-            {data.showSteps ? "Close Steps" : "Change My Rate"}
+            {seed.showSteps ? "Close Steps" : "Change My Rate"}
           </Button>
         </Grid>
         <Grid item xs={12}>
-          {data.showSteps && renderMixRateSteps(data)}
+          {seed.showSteps && (
+            <MixRatioSteps
+              seed={seed}
+              council={council}
+              renderFormLabel={renderFormLabel}
+              updateSeed={updateSeed}
+              generatePercentInGroup={generatePercentInGroup}
+            />
+          )}
         </Grid>
       </Grid>
     );
@@ -321,365 +326,7 @@ const MixRatio = ({ council }) => {
     console.log("count...", group, count);
     return 1 / count;
   };
-  const renderMixRateSteps = (seed) => {
-    const percentInGroup = generatePercentInGroup(seed);
-    return (
-      <Grid container xs={12}>
-        {/* NECCC Step 1:  */}
-        {council === "NECCC" && (
-          <>
-            <Grid item xs={12}>
-              <Typography className="mix-ratio-step-header">
-                Step 1:{" "}
-              </Typography>
-            </Grid>
-            {renderFormLabel(
-              "Mix Seeding Rate PLS",
-              "% in Group",
-              "% of Single Species Seeding Rate"
-            )}
-            <Grid item xs={3}>
-              <NumberTextField
-                className="text-field-100"
-                id="filled-basic"
-                disabled={false}
-                label="Single Species Seeding Rate PLS"
-                variant="filled"
-                handleChange={(e) => {
-                  updateSeed(
-                    e.target.value,
-                    "singleSpeciesSeedingRatePLS",
-                    seed
-                  );
-                }}
-                value={seed.singleSpeciesSeedingRatePLS}
-              />
-              <Typography>Lbs / Acre</Typography>
-            </Grid>
 
-            <Grid item xs={1}>
-              <Typography className="math-icon">x</Typography>
-            </Grid>
-            <Grid item xs={3}>
-              <NumberTextField
-                className="text-field-100"
-                id="filled-basic"
-                label="% in Group"
-                variant="filled"
-                disabled={true}
-                value={Math.round(convertToPercent(percentInGroup))}
-              />
-            </Grid>
-            <Grid item xs={1}>
-              <Typography className="math-icon">x</Typography>
-            </Grid>
-            <Grid item xs={3}>
-              <NumberTextField
-                className="text-field-100"
-                id="filled-basic"
-                label="% of Single Species Rate"
-                variant="filled"
-                disabled={false}
-                handleChange={(e) => {
-                  updateSeed(
-                    e.target.value,
-                    "percentOfSingleSpeciesRate",
-                    seed
-                  );
-                }}
-                value={Math.round(seed.percentOfSingleSpeciesRate)}
-              />
-              <Typography>{council === "MCCC" ? "MCCC" : "NECCC"}</Typography>
-            </Grid>
-            <Grid container className="steps-row-2" xs={12}>
-              <Grid item xs={4}>
-                <Typography className="math-icon">=</Typography>
-              </Grid>
-              <Grid item xs={7}>
-                <NumberTextField
-                  className="text-field-100"
-                  id="filled-basic"
-                  label="Mix Seeding Rate"
-                  disabled={true}
-                  variant="filled"
-                  value={seed.mixSeedingRate}
-                />
-                <Typography>Lbs / Acre</Typography>
-              </Grid>
-              <Grid item xs={1}></Grid>
-            </Grid>
-          </>
-        )}
-        {/* NECCC Step 1 END */}
-
-        {/* MCCC Step 1 START */}
-        {council === "MCCC" && (
-          <>
-            <Grid item xs={12}>
-              <Typography className="mix-ratio-step-header">Step 1:</Typography>
-            </Grid>
-            {renderFormLabel(
-              "Single Species Seeding Rate PLS",
-              "% of Single Species Rate",
-              "Mix Seeding Rate"
-            )}
-            <Grid container xs={12} className="mix-ratio-form-container">
-              <Grid item xs={3}>
-                <NumberTextField
-                  className="text-field-100"
-                  id="filled-basic"
-                  disabled={false}
-                  label="Single Species Seeding Rate PLS"
-                  variant="filled"
-                  handleChange={(e) => {
-                    updateSeed(
-                      e.target.value,
-                      "singleSpeciesSeedingRatePLS",
-                      seed
-                    );
-                  }}
-                  value={seed.singleSpeciesSeedingRatePLS}
-                />
-                <Typography>Lbs / Acre</Typography>
-              </Grid>
-              <Grid item xs={1}>
-                <Typography className="math-icon">X</Typography>
-              </Grid>
-              <Grid item xs={3}>
-                <NumberTextField
-                  className="text-field-100"
-                  id="filled-basic"
-                  label="% of Single Species Rate"
-                  variant="filled"
-                  disabled={false}
-                  handleChange={(e) => {
-                    updateSeed(
-                      e.target.value,
-                      "percentOfSingleSpeciesRate",
-                      seed
-                    );
-                  }}
-                  value={seed.percentOfSingleSpeciesRate}
-                />
-                <Typography>{council === "MCCC" ? "MCCC" : "NECCC"}</Typography>
-              </Grid>
-              <Grid item xs={1}>
-                <Typography className="math-icon">=</Typography>
-              </Grid>
-              <Grid item xs={3}>
-                <NumberTextField
-                  className="text-field-100"
-                  id="filled-basic"
-                  label="Mix Seeding Rate"
-                  disabled={true}
-                  variant="filled"
-                  value={seed.mixSeedingRate}
-                />
-                <Typography>Lbs / Acre</Typography>
-              </Grid>
-            </Grid>
-          </>
-        )}
-
-        {/* MCCC Step 1 END */}
-        <Grid item xs={12}>
-          <Typography className="mix-ratio-step-header">Step 2: </Typography>
-        </Grid>
-        {renderFormLabel(
-          "Single Species Seeding Rate PLS",
-          "% of Single Species Rate",
-          "Mix Seeding Rate"
-        )}
-        <Grid item xs={3}>
-          <NumberTextField
-            className="text-field-100"
-            id="filled-basic"
-            disabled={true}
-            label="Seeds / Pound"
-            variant="filled"
-            handleChange={(e) => {
-              updateSeed(e.target.value, "seedsPerPound", seed);
-            }}
-            value={seed.seedsPerPound}
-          />
-        </Grid>
-        <Grid item xs={1}>
-          <Typography className="math-icon">X</Typography>
-        </Grid>
-        <Grid item xs={3}>
-          <NumberTextField
-            className="text-field-100"
-            id="filled-basic"
-            disabled={true}
-            label="Mix Seeding Rate"
-            variant="filled"
-            value={seed.mixSeedingRate}
-          />
-          <Typography>Lbs / Acre</Typography>
-        </Grid>
-        <Grid item xs={1}>
-          <Typography className="math-icon">=</Typography>
-        </Grid>
-        <Grid item xs={3}>
-          <NumberTextField
-            className="text-field-100"
-            id="filled-basic"
-            label="Seeds / Acre"
-            variant="filled"
-            disabled={true}
-            value={seed.seedsPerAcre}
-          />
-        </Grid>
-        {council === "NECCC" && (
-          <>
-            {" "}
-            <Grid item xs={12}>
-              <Typography className="mix-ratio-step-header">
-                Step 3:{" "}
-              </Typography>
-            </Grid>
-            {renderFormLabel("Seeds/Acre", "Sq. Ft. / Acres", "Plants/Acre")}
-            <Grid item xs={3}>
-              <NumberTextField
-                className="text-field-100"
-                id="filled-basic"
-                label="Seeds / Acre"
-                variant="filled"
-                disabled={true}
-                value={seed.seedsPerAcre}
-              />
-            </Grid>
-            <Grid item xs={1}>
-              <Typography className="math-icon">÷</Typography>
-            </Grid>
-            <Grid item xs={3}>
-              <NumberTextField
-                className="text-field-100"
-                id="filled-basic"
-                label="Sq. Ft./ Acre"
-                variant="filled"
-                disabled={true}
-                value={seed.sqFtAcre}
-              />
-            </Grid>
-            <Grid item xs={1}>
-              <Typography className="math-icon">=</Typography>
-            </Grid>
-            <Grid item xs={3}>
-              <NumberTextField
-                className="text-field-100"
-                id="filled-basic"
-                label="Seeds / Sq. Ft."
-                variant="filled"
-                disabled={true}
-                value={seed.aproxPlantsSqFt}
-              />
-            </Grid>
-          </>
-        )}
-        {council === "MCCC" && (
-          <>
-            <Grid item xs={12}>
-              <Typography className="mix-ratio-step-header">
-                Step 3:{" "}
-              </Typography>
-            </Grid>
-            {renderFormLabel("Seeds/Acre", "% Survival", "Plants/Acre")}
-            <Grid item xs={3}>
-              <NumberTextField
-                className="text-field-100"
-                id="filled-basic"
-                label="Seeds / Acre"
-                variant="filled"
-                disabled={true}
-                value={seed.seedsPerAcre}
-              />
-            </Grid>
-            <Grid item xs={1}>
-              <Typography className="math-icon">X</Typography>
-            </Grid>
-            <Grid item xs={3}>
-              <NumberTextField
-                className="text-field-100"
-                id="filled-basic"
-                label="% Survival"
-                variant="filled"
-                disabled={false}
-                handleChange={(e) => {
-                  updateSeed(
-                    convertToDecimal(e.target.value),
-                    "percentChanceOfWinterSurvival",
-                    seed
-                  );
-                }}
-                value={convertToPercent(seed.percentChanceOfWinterSurvival)}
-              />
-            </Grid>
-            <Grid item xs={1}>
-              <Typography className="math-icon">=</Typography>
-            </Grid>
-            <Grid item xs={3}>
-              <NumberTextField
-                className="text-field-100"
-                id="filled-basic"
-                label="Plants / Acre"
-                variant="filled"
-                disabled={true}
-                value={seed.plantsPerAcre}
-              />
-            </Grid>
-            <Grid item xs={12}>
-              <Typography className="mix-ratio-step-header">
-                Step 4:{" "}
-              </Typography>
-            </Grid>
-            {renderFormLabel(
-              "Plants/Acre",
-              "Sq.Ft./Acre",
-              "Aproximate Plants/Sq.Ft."
-            )}
-            <Grid item xs={3}>
-              <NumberTextField
-                className="text-field-100"
-                id="filled-basic"
-                label="Plants / Acre"
-                variant="filled"
-                disabled={true}
-                value={seed.plantsPerAcre}
-              />
-            </Grid>
-            <Grid item xs={1}>
-              <Typography className="math-icon">÷</Typography>
-            </Grid>
-            <Grid item xs={3}>
-              <NumberTextField
-                className="text-field-100"
-                id="filled-basic"
-                label="Sq. Ft./ Acre"
-                variant="filled"
-                disabled={true}
-                value={seed.sqFtAcre}
-              />
-            </Grid>
-            <Grid item xs={1}>
-              <Typography className="math-icon">=</Typography>
-            </Grid>
-            <Grid item xs={3}>
-              <NumberTextField
-                className="text-field-100"
-                id="filled-basic"
-                label="Aproximate Plants  / Sq. Ft."
-                variant="filled"
-                disabled={true}
-                value={seed.aproxPlantsSqFt}
-              />
-              <Typography>Lbs / Acre</Typography>
-            </Grid>
-          </>
-        )}
-      </Grid>
-    );
-  };
   return (
     <Grid xs={12} container>
       {renderSeedsSelected()}
